feat: Restructure Korge project for unified application namespace and entry points
Try to get jvmRun working in "korge-sandbox".
