+++ korge-sandbox/androidApp/src/main/kotlin/org/korge/application/MainActivity.kt
@@ -1,6 +1,4 @@
-import korlibs.render.GameWindowCreationConfig
-import korlibs.render.KorgwActivity
-import org.korge.sandbox.main
+package org.korge.application
 
 class MainActivity : KorgwActivity(config = GameWindowCreationConfig(msaa = 1)) {
     override suspend fun activityMain() {
