+++ korge-sandbox/shared/src/commonMain/kotlin/org/korge/application/helpers.kt
@@ -1,4 +1,4 @@
-package org.korge.sandbox
+package org.korge.application
 
 import korlibs.io.async.*
 import korlibs.io.lang.*
