--- korge-sandbox/webApp/src/commonMain/kotlin/main.kt
@@ -1,8 +1,0 @@
-package org.korge.sandbox.web
-
-import org.korge.sandbox.main as sandboxMain
-
-// Web (JS / WASM) entry point. Delegates to the shared sandbox game entry.
-suspend fun main() {
-    sandboxMain()
-}
