+++ korge-sandbox/webApp/src/commonMain/kotlin/org/korge/application/web/main.kt
@@ -1,0 +1,8 @@
+package org.korge.application.web
+
+import org.korge.application.main as applicationMain
+
+// Web (JS / WASM) entry point. Delegates to the shared application game entry.
+suspend fun main() {
+    applicationMain()
+}
